Moved InforOperation and WSHubOperation to separate classes
